+++ lib/vcs_base/src/err.ml
@@ -20,4 +20,12 @@
 (*******************************************************************************)
 
 include Vcs.Err
-include Vcs.Err.Private.Vcs_base
+
+let to_error t =
+  Error.create_s
+    (match Private.view t with
+     | { steps = []; error } -> error
+     | { steps = _ :: _; error = _ } -> sexp_of_t t)
+;;
+
+let of_error error = create_s (Error.sexp_of_t error)