+++ lib/vcs_base/src/vcs_or_error0.mli
@@ -19,45 +19,4 @@
 (*_  <http://www.gnu.org/licenses/> and <https://spdx.org>, respectively.       *)
 (*_******************************************************************************)
 
-(** Manipulating the output of process run by vcs and providers - typically the
-    ["git"] command. *)
-
-module Output : sig
-  type t = Git_output0.t =
-    { exit_code : int
-    ; stdout : string
-    ; stderr : string
-    }
-  [@@deriving sexp_of]
-end
-
-(** This is the interface commonly used by raising and non-raising helper
-    modules, such as {!module:Vcs.Git}, {!module:Vcs_base.Vcs.Git.Or_error},
-    {!module:Vcs.Git.Result}, {!module:Vcs.Git.Rresult}, and custom ones built
-    with {!module:Vcs.Git.Non_raising.Make}. [S] is parametrized by the result
-    type returned by the helpers. *)
-module type S = Vcs_interface.Process_S
-
-(** The interface exposed at the top level of this module are helpers that
-    return direct results, or raise {!exception:Vcs.E}. This module is
-    exported to users as [Vcs.Git].
-
-    The helpers are suitable for use in combination with the {!val:Vcs.git}
-    function, which will take care of wrapping the exception with useful
-    context, before re-raising it. *)
-include S with type 'a result := 'a
-
-module Non_raising : sig
-  (** A functor to build non raising helpers based on a custom error type.
-
-      In addition to {!module:Vcs.Git.Result}, {!module:Vcs.Git.Rresult} and
-      {!module:Vcs_base.Vcs.Git.Or_error}, we provide this functor to create a
-      [Git.S] interface based on a custom error type of your choice. *)
-
-  module type M = Vcs_interface.Error_S
-
-  module Make (M : M) : S with type 'a result := ('a, M.t) Result.t
-end
-
-module Rresult : S with type 'a result := ('a, Vcs_rresult0.t) Result.t
-module Result : S with type 'a result := ('a, Err.t) Result.t
+include Vcs.Non_raising.M with type t = Error.t